[DEV-7216] Styling for COVID Law filter
