+++ src/_scss/layouts/default/header/header.scss
@@ -8,20 +8,3 @@
     }
 }
 
-.usda-page-header {
-    @import "../../../mixins/fullSectionWrap.scss";
-    .usda-page-header__container {
-        .usda-page-header__header,
-        .usda-page-header__toolbar {
-            @media (max-width: $medium-screen) {
-                .usda-fy-picker__container {
-                    padding-left: 0;
-                }
-                padding-left: $outer-gutter / 2;
-            }
-        }
-        @media(min-width: $medium-screen) {
-            @include fullSectionWrap(0, 0);
-        }
-    }
-}
