+++ src/js/components/covid19/Covid19Page.jsx
@@ -3,8 +3,6 @@
  * Created by Jonathan Hill 06/02/20
  */
 
-// TODO: DEV-7122 Move to new Page Header Component
-
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { useDispatch, useSelector } from 'react-redux';
@@ -38,30 +36,9 @@ import Analytics from 'helpers/analytics/Analytics';
 import { componentByCovid19Section } from 'containers/covid19/helpers/covid19';
 import DownloadButtonContainer from 'containers/covid19/DownloadButtonContainer';
 import SidebarFooter from 'components/covid19/SidebarFooter';
+import PublicLawPicker from './PublicLawPicker';
 
 require('pages/covid19/index.scss');
-
-const options = [
-    { title: 'All Related Public Laws', description: 'All data related to DEFC L, M, O, U, and V.', value: 'all' },
-    { title: 'American Rescue Plan Act of 2021', description: 'Emergency PublicLaw 117-7 (DEFC V)', value: 'biden' },
-    { title: 'Learn more about filteringUSAspending data by Public Law in our Data Sources & Methodology page.', value: 'dsm' }
-];
-
-const PublicLawPicker = ({
-    selectedOption = 'all'
-}) => {
-    return (
-        <div className="public-law-picker__container">
-            <Picker
-                className="public-law-picker"
-                icon={<FontAwesomeIcon icon="scroll" size="lg" />}
-                selectedOption={options[0].title}
-                options={options.map((obj) => ({ name: obj.title, value: obj.value }))} />
-            <span>Public Law Filter</span>
-        </div>
-    );
-};
-
 
 const propTypes = {
     areDefCodesLoading: PropTypes.bool
